+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ CHANGELOG
 Unreleased
 ==========
 
+* Added: Support for reading vulns and asset exports into memory
 * Added: Support for graceful back-off
 
 1.4.1